requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
Jinja2==3.1.2
openai==1.3.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
python-multipart==0.0.6
requests==2.31.0
Pillow==10.1.0